% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_console}
\alias{appender_console}
\alias{appender_stderr}
\title{Append log record to stderr}
\usage{
appender_console(lines)

appender_stderr(lines)
}
\arguments{
\item{lines}{character vector}
}
\description{
Append log record to stderr
}
\seealso{
Other log_appenders:
\code{\link[=appender_async]{appender_async()}},
\code{\link[=appender_file]{appender_file()}},
\code{\link[=appender_kinesis]{appender_kinesis()}},
\code{\link[=appender_ntfy]{appender_ntfy()}},
\code{\link[=appender_pushbullet]{appender_pushbullet()}},
\code{\link[=appender_slack]{appender_slack()}},
\code{\link[=appender_stdout]{appender_stdout()}},
\code{\link[=appender_syslog]{appender_syslog()}},
\code{\link[=appender_tee]{appender_tee()}},
\code{\link[=appender_telegram]{appender_telegram()}}
}
\concept{log_appenders}
