% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_ms_teams}
\alias{appender_ms_teams}
\title{Send log messages to Microsoft Teams Incoming Webhook}
\usage{
appender_ms_teams(teams_webhook_url)
}
\arguments{
\item{teams_webhook_url}{character, url to the incoming webhook you would like to use

You must create an incoming webhook for the Microsoft Teams channel you want to send log messages
to. For instructions on how to do this, visit the \href{https://docs.microsoft.com/en-us/microsoftteams/platform/webhooks-and-connectors/how-to/add-incoming-webhook#add-an-incoming-webhook-to-a-teams-channel}{Microsoft Teams documentation page}.}
}
\value{
function taking lines as argument
}
\description{
Send log messages to Microsoft Teams Incoming Webhook
}
\note{
This functionality depends on the \pkg{httr} package.
}
\seealso{
This is generator function for \code{\link{log_appender}}, for alternatives, see eg \code{\link{appender_console}}, \code{\link{appender_file}}, \code{\link{appender_tee}}, \code{\link{appender_slack}}, \code{\link{appender_telegram}}, \code{\link{appender_syslog}}, \code{\link{appender_kinesis}} and \code{\link{appender_async}} for evaluate any \code{\link{log_appender}} function in a background process.
}
