% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_ntfy}
\alias{appender_ntfy}
\title{Send log messages to ntfy}
\usage{
appender_ntfy(title = "{logger}", tags = c("memo"), ...)
}
\arguments{
\item{title}{notification title}

\item{tags}{emoji (or general tag) for notification. See \link[ntfy:emoji]{ntfy::emoji}}

\item{...}{extra parameters passed to \link[ntfy:ntfy_send]{ntfy::ntfy_send} such as
\code{priority}, \code{topic}, etc.}
}
\description{
Send log messages to ntfy
}
\details{
Configure server and topic via environment variables. See
\code{\link[ntfy:ntfy_topic]{ntfy::ntfy_topic()}} for details
}
\note{
This functionality depends on the \pkg{ntfy} package.
}
\seealso{
Other log_appenders:
\code{\link[=appender_async]{appender_async()}},
\code{\link[=appender_console]{appender_console()}},
\code{\link[=appender_file]{appender_file()}},
\code{\link[=appender_kinesis]{appender_kinesis()}},
\code{\link[=appender_pushbullet]{appender_pushbullet()}},
\code{\link[=appender_slack]{appender_slack()}},
\code{\link[=appender_stdout]{appender_stdout()}},
\code{\link[=appender_syslog]{appender_syslog()}},
\code{\link[=appender_tee]{appender_tee()}},
\code{\link[=appender_telegram]{appender_telegram()}}
}
\concept{log_appenders}
