% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_slack}
\alias{appender_slack}
\title{Send log messages to a Slack channel}
\usage{
appender_slack(
  channel = Sys.getenv("SLACK_CHANNEL"),
  username = Sys.getenv("SLACK_USERNAME"),
  icon_emoji = Sys.getenv("SLACK_ICON_EMOJI"),
  api_token = Sys.getenv("SLACK_API_TOKEN"),
  preformatted = TRUE
)
}
\arguments{
\item{channel}{Slack channel name with a hashtag prefix for public
channel and no prefix for private channels}

\item{username}{Slack (bot) username}

\item{icon_emoji}{optional override for the bot icon}

\item{api_token}{Slack API token}

\item{preformatted}{use code tags around the message?}
}
\value{
function taking \code{lines} argument
}
\description{
Send log messages to a Slack channel
}
\note{
This functionality depends on the \pkg{slackr} package.
}
\seealso{
Other log_appenders:
\code{\link[=appender_async]{appender_async()}},
\code{\link[=appender_console]{appender_console()}},
\code{\link[=appender_file]{appender_file()}},
\code{\link[=appender_kinesis]{appender_kinesis()}},
\code{\link[=appender_ntfy]{appender_ntfy()}},
\code{\link[=appender_pushbullet]{appender_pushbullet()}},
\code{\link[=appender_stdout]{appender_stdout()}},
\code{\link[=appender_syslog]{appender_syslog()}},
\code{\link[=appender_tee]{appender_tee()}},
\code{\link[=appender_telegram]{appender_telegram()}}
}
\concept{log_appenders}
