% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_syslognet}
\alias{appender_syslognet}
\title{Send log messages to a network syslog server}
\usage{
appender_syslognet(identifier, server, port = 601L)
}
\arguments{
\item{identifier}{program/function identification (string).}

\item{server}{machine where syslog daemon runs (string).}

\item{port}{port where syslog daemon listens (integer).}
}
\value{
A function taking a \code{lines} argument.
}
\description{
Send log messages to a network syslog server
}
\note{
This functionality depends on the \pkg{syslognet} package.
}
\examples{
\dontrun{
if (requireNamespace("syslognet", quietly = TRUE)) {
  log_appender(appender_syslognet("test_app", "remoteserver"))
  log_info("Test message.")
}
}
}
