% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_telegram}
\alias{appender_telegram}
\title{Send log messages to a Telegram chat}
\usage{
appender_telegram(
  chat_id = Sys.getenv("TELEGRAM_CHAT_ID"),
  bot_token = Sys.getenv("TELEGRAM_BOT_TOKEN"),
  parse_mode = NULL
)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of
the target channel (in the format @channelusername)}

\item{bot_token}{Telegram Authorization token}

\item{parse_mode}{Message parse mode. Allowed values: Markdown or
HTML}
}
\value{
function taking \code{lines} argument
}
\description{
Send log messages to a Telegram chat
}
\note{
This functionality depends on the \pkg{telegram} package.
}
\seealso{
Other log_appenders:
\code{\link[=appender_async]{appender_async()}},
\code{\link[=appender_console]{appender_console()}},
\code{\link[=appender_file]{appender_file()}},
\code{\link[=appender_kinesis]{appender_kinesis()}},
\code{\link[=appender_ntfy]{appender_ntfy()}},
\code{\link[=appender_pushbullet]{appender_pushbullet()}},
\code{\link[=appender_slack]{appender_slack()}},
\code{\link[=appender_stdout]{appender_stdout()}},
\code{\link[=appender_syslog]{appender_syslog()}},
\code{\link[=appender_tee]{appender_tee()}}
}
\concept{log_appenders}
