% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_void}
\alias{appender_void}
\title{Dummy appender not delivering the log record to anywhere}
\usage{
appender_void(lines)
}
\arguments{
\item{lines}{character vector}
}
\description{
Dummy appender not delivering the log record to anywhere
}
