% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{as.character.logger}
\alias{as.character.logger}
\title{Convert a log message to character}
\usage{
\method{as.character}{logger}(
  x,
  ...,
  namespace = x$namespace,
  index = 1,
  .topenv = parent.frame()
)
}
\arguments{
\item{namespace}{string}

\item{index}{index of the logger within the namespace}
}
\description{
Convert a log message to character
}
