% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{as.loglevel}
\alias{as.loglevel}
\title{Convert R object into a logger log-level}
\usage{
as.loglevel(x)
}
\arguments{
\item{x}{string or integer}
}
\value{
pander log-level, e.g. \code{INFO}
}
\description{
Convert R object into a logger log-level
}
\examples{
as.loglevel(INFO)
as.loglevel(400L)
as.loglevel(400)
}
