% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catch_base_log}
\alias{catch_base_log}
\title{Catch the log header}
\usage{
catch_base_log(
  level,
  namespace,
  .topcall = sys.call(-1),
  .topenv = parent.frame()
)
}
\arguments{
\item{level}{see \code{\link{log_levels}}}

\item{namespace}{string}
}
\value{
string
}
\description{
Catch the log header
}
\examples{
\dontrun{
catch_base_log(INFO, NA_character_)
logger <- layout_glue_generator(format = '{node}/{pid}/{namespace}/{fn} {time} {level}: {msg}')
log_layout(logger)
catch_base_log(INFO, NA_character_)
fun <- function() catch_base_log(INFO, NA_character_)
fun()
catch_base_log(INFO, NA_character_, .topcall = call("funLONG"))
}
}
\keyword{internal}
