% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerinfo.R
\name{check_datetime_cols}
\alias{check_datetime_cols}
\title{Checks for datetime columns}
\usage{
check_datetime_cols(data, column)
}
\arguments{
\item{data}{The dataframe in which the datetime columns will be checked}

\item{column}{The name of the datetime column}
}
\value{
Returns a `rlang::abort()` message in case the datetime values are not correctly formatted
}
\description{
Helper function to check if a `column` of a `data`frame is formatted as POSIXct
}
\examples{
\dontrun{
check_datetime_cols(loggerinfo, "retrieved")
}

}
