% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerinfo.R
\name{check_NA_cols}
\alias{check_NA_cols}
\title{Check NA in columns}
\usage{
check_NA_cols(data, column)
}
\arguments{
\item{data}{The dataframe in which for specified columns will be checked if NA values occur}

\item{column}{The column that should be checked}
}
\value{
Returns a `rlang::abort()` message in case NA values occur in the selected column
}
\description{
Helper function to check if a `column` of a `data`frame contains NA values
}
\examples{
\dontrun{
check_NA_cols(loggerinfo, "retrieved")
}

}
