% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerfiles.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Clean logger column names}
\usage{
clean_colnames(data, rownumber)
}
\arguments{
\item{data}{The current logger csv}

\item{rownumber}{The row number in the `loggerinfo` dataframe corresponding to `data`.}
}
\value{
a character vector of clean column names.
}
\description{
The function extracts the column names from the logger csv files and cleans them.
These names contain special characters, units, and the serial number and will be standardized to date_time, temperature, and light_intensity.
}
\examples{
\dontrun{
names <- clean_colnames(data, row_i)
}

}
