% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2spatial.R
\name{convert2spatial}
\alias{convert2spatial}
\title{\code{convert2spatial} Convert selected parts of sightings data list to shapefiles.}
\usage{
convert2spatial(
  data.list = midnat,
  zone = 22,
  hemisphere = "south",
  whales = T,
  birds = T,
  shp.dir = "./shapefiles"
)
}
\arguments{
\item{data.list}{List of data frames created using \code{readLogger}}

\item{zone}{Which UTM zone should data be converted to.
If this is NA (default), then shapefiles are in geographic (Lat/Lon) projection.
To convert to UTM, set this argument to desired zone (22 for the Scotia Sea).}

\item{hemisphere}{Hemisphere where data were collected.}

\item{whales}{Logical, indicating if whale sightings table should be read}

\item{birds}{Logical, indicating if bird sightings tables should be read}

\item{shp.dir}{Directory to which output shapefiles should be written.}
}
\value{
Writes shapefiles to specified directory.
}
\description{
\code{convert2spatial} Convert selected parts of sightings data list to shapefiles.
}
\details{
Writes ESRI shapefiles that can be read by standard GIS software.
}
\seealso{
\code{\link{readLoggerTable}} to read single table from sightings database,
  \code{\link{readLogger}} to read entire sightings database,
  \code{\link{correctEffort}} tto correct inconsistencies in effort table,
  \code{\link{filterGPS}} to flag bad GPS positions,
  \code{\link{fillGPS}} To fill in missing GPS positions after filtering,
  \code{\link{mkLeaflet}} to make interactive Leaflet map of sightings.

Other Logger sightings database functions: 
\code{\link{correctEffort}()},
\code{\link{fillGPS}()},
\code{\link{filterGPS}()},
\code{\link{mkLeaflet}()},
\code{\link{readLoggerTable}()},
\code{\link{readLogger}()}
}
\author{
Martin Biuw
}
\concept{Logger sightings database functions}
