% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctEffort.R
\name{correctEffort}
\alias{correctEffort}
\title{\code{correctEffort} Correct errors start & end times in effort table}
\usage{
correctEffort(
  data.list = midnat,
  include.birds = F,
  tGap = 60,
  plotType = "dygraph"
)
}
\arguments{
\item{data.list}{List of data frames created using \code{readLogger}}

\item{include.birds}{Logical, indicating if bird sightings table should be used
when correcting effort. Default is FALSE, i.e. only whale sightings and effort tables are used.}

\item{tGap}{Time (in minutes) at which new effort start is assumed even if first data record
does not have a designated start ('SE') event}
}
\value{
Returns the same list as in the input, but with corrected start and end of effort designations,
and with all sightings and gps points allocated to transects (defined as contiguous effort periods).
}
\description{
\code{correctEffort} Correct errors start & end times in effort table
}
\seealso{
\code{\link{readLoggerTable}} to read single table from sightings database,
  \code{\link{readLogger}} to read entire sightings database,
  \code{\link{filterGPS}} to flag bad GPS positions,
  \code{\link{fillGPS}} to fill in missing GPS positions after filtering
  \code{\link{mkLeaflet}} to make interactive Leaflet map of sightings,
  \code{\link{convert2spatial}} to convert selected parts of sightings data list to shapefiles.

Other Logger sightings database functions: 
\code{\link{convert2spatial}()},
\code{\link{fillGPS}()},
\code{\link{filterGPS}()},
\code{\link{mkLeaflet}()},
\code{\link{readLoggerTable}()},
\code{\link{readLogger}()}
}
\author{
Martin Biuw
}
\concept{Logger sightings database functions}
