% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespaces.R
\name{delete_logger_index}
\alias{delete_logger_index}
\title{Delete an index from a logger namespace}
\usage{
delete_logger_index(namespace = "global", index)
}
\arguments{
\item{namespace}{logger namespace}

\item{index}{index of the logger within the namespace}
}
\description{
Delete an index from a logger namespace
}
