% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_to_one_line}
\alias{deparse_to_one_line}
\title{Deparse and join all lines into a single line}
\usage{
deparse_to_one_line(x)
}
\arguments{
\item{x}{object to \code{deparse}}
}
\value{
string
}
\description{
Calling \code{deparse} and joining all the returned lines into a
single line, separated by whitespace, and then cleaning up all the
duplicated whitespace (except for excessive whitespace in strings
between single or double quotes).
}
