% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{fallback_namespace}
\alias{fallback_namespace}
\title{Checks if provided namespace exists and falls back to global if not}
\usage{
fallback_namespace(namespace)
}
\arguments{
\item{namespace}{string}
}
\value{
string
}
\description{
Checks if provided namespace exists and falls back to global if not
}
\keyword{internal}
