% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterGPS.R
\name{filterGPS}
\alias{filterGPS}
\title{\code{filterGPS} Filter out bad GPS position estimates}
\usage{
filterGPS(data.list = midnat, maxSpeed = 300)
}
\arguments{
\item{data.list}{List of data frames created using \code{readLogger}}

\item{maxSpeed}{Maximum speed allowed between points, MORE DETAILS NEEDED HERE!!}
}
\value{
Returns the same list as in the input, but with flagged GPS positions in the gps table,
}
\description{
\code{filterGPS} Filter out bad GPS position estimates
}
\details{
Creates a quality 'flag' variable in the gps table, indicating which points have passed the
quality criterion and should be considered for further processing and analyses.
}
\seealso{
\code{\link{readLoggerTable}} to read single table from sightings database,
  \code{\link{readLogger}} to read entire sightings database,
  \code{\link{correctEffort}} to correct inconsistencies in effort table,
  \code{\link{fillGPS}} To fill in missing GPS positions after filtering
  \code{\link{mkLeaflet}} to make interactive Leaflet map of sightings,
  \code{\link{convert2spatial}} to convert selected parts of sightings data list to shapefiles.

Other Logger sightings database functions: 
\code{\link{convert2spatial}()},
\code{\link{correctEffort}()},
\code{\link{fillGPS}()},
\code{\link{mkLeaflet}()},
\code{\link{readLoggerTable}()},
\code{\link{readLogger}()}
}
\author{
Martin Biuw
}
\concept{Logger sightings database functions}
