% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{find_call}
\alias{find_call}
\title{Find the call from which the logging function was called}
\usage{
find_call()
}
\value{
string
}
\description{
Find the call from which the logging function was called
}
\keyword{internal}
