% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{find_fn}
\alias{find_fn}
\title{Find the function from which the logging function was called}
\usage{
find_fn()
}
\value{
string
}
\description{
Find the function from which the logging function was called
}
\keyword{internal}
