% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{find_namespace}
\alias{find_namespace}
\title{Find the namespace from which the logging function was called}
\usage{
find_namespace()
}
\value{
string
}
\description{
Find the namespace from which the logging function was called
}
\keyword{internal}
