% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R
\name{find_parents}
\alias{find_parents}
\title{Find the namespace, function name and call from which the logging function was called}
\usage{
find_parents()
}
\value{
list
}
\description{
Find the namespace, function name and call from which the logging function was called
}
\keyword{internal}
