% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_glue}
\alias{formatter_glue}
\title{Apply \code{\link[glue:glue]{glue::glue()}} to convert R objects into a character vector}
\usage{
formatter_glue(
  ...,
  .logcall = sys.call(),
  .topcall = sys.call(-1),
  .topenv = parent.frame()
)
}
\arguments{
\item{...}{passed to \code{\link[glue:glue]{glue::glue()}} for the text interpolation}

\item{.logcall}{the logging call being evaluated (useful in
formatters and layouts when you want to have access to the raw,
unevaluated R expression)}

\item{.topcall}{R expression from which the logging function was
called (useful in formatters and layouts to extract the calling
function's name or arguments)}

\item{.topenv}{original frame of the \code{.topcall} calling function
where the formatter function will be evaluated and that is used
to look up the \code{namespace} as well via \code{logger:::top_env_name}}
}
\value{
character vector
}
\description{
Apply \code{\link[glue:glue]{glue::glue()}} to convert R objects into a character vector
}
\note{
Although this is the default log message formatter function,
but when \pkg{glue} is not installed, \code{\link[=formatter_sprintf]{formatter_sprintf()}}
will be used as a fallback.
}
\seealso{
Other log_formatters:
\code{\link[=formatter_cli]{formatter_cli()}},
\code{\link[=formatter_glue_or_sprintf]{formatter_glue_or_sprintf()}},
\code{\link[=formatter_glue_safe]{formatter_glue_safe()}},
\code{\link[=formatter_json]{formatter_json()}},
\code{\link[=formatter_logging]{formatter_logging()}},
\code{\link[=formatter_pander]{formatter_pander()}},
\code{\link[=formatter_paste]{formatter_paste()}},
\code{\link[=formatter_sprintf]{formatter_sprintf()}}
}
\concept{log_formatters}
