% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_glue_or_sprinf}
\alias{formatter_glue_or_sprinf}
\title{Apply \code{glue} and \code{sprintf}}
\usage{
formatter_glue_or_sprinf(msg, ...)
}
\arguments{
\item{...}{passed to \code{glue} for the text interpolation}
}
\value{
character vector
}
\description{
Apply \code{glue} and \code{sprintf}
}
