% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_logger}
\alias{get_logger}
\title{Find the logger used in the current namespace}
\usage{
get_logger()
}
\value{
function
}
\description{
Find the logger used in the current namespace
}
\keyword{internal}
