% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerfiles.R
\name{get_SN}
\alias{get_SN}
\title{Get serial number from logger csv files}
\usage{
get_SN(data, sep = ".")
}
\arguments{
\item{data}{The current logger csv}

\item{sep}{The separating character between the parameter name and the serial number. Defaults to `.`}
}
\value{
The serial number.
}
\description{
The serial number is stored in the column names if the data was exported correctly as csv.
This function extracts this serial number so it can be compared to the one provided in the logger metadata.
}
\examples{
\dontrun{
SN <- get_SN(data)
}

}
