\name{getLoggingThreshold}
\alias{getLoggingThreshold}
\title{getter for logging threshold}
\usage{
  getLoggingThreshold()
}
\value{
  integer value for the logging threshold
}
\description{
  getter function for logging threshold. Maximum header
  level which will be logged, default value is 10.
}
\author{
  Dr Rainer M Krug
}

