% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{grayscale_by_log_level}
\alias{grayscale_by_log_level}
\title{Render a string with light/dark gray based on the related log level}
\usage{
grayscale_by_log_level(msg, level)
}
\arguments{
\item{msg}{string}

\item{level}{see \code{\link{log_levels}}}
}
\value{
string with ANSI escape code
}
\description{
Adding color to a string to be used in terminal output. Supports ANSI standard colors 8 or 256.
}
\examples{
\dontshow{if (requireNamespace("crayon")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cat(grayscale_by_log_level('foobar', FATAL), '\n')
cat(grayscale_by_log_level('foobar', ERROR), '\n')
cat(grayscale_by_log_level('foobar', WARN), '\n')
cat(grayscale_by_log_level('foobar', SUCCESS), '\n')
cat(grayscale_by_log_level('foobar', INFO), '\n')
cat(grayscale_by_log_level('foobar', DEBUG), '\n')
cat(grayscale_by_log_level('foobar', TRACE), '\n')
\dontshow{\}) # examplesIf}
}
