\name{incHeaderLevel}
\alias{.incHeaderLevel}
\alias{incHeaderLevel}
\title{incHeaderLevel}
\usage{
  .incHeaderLevel(header, clock)
}
\arguments{
  \item{header}{character string, Header of the new header
  level}

  \item{clock}{character string, containing the timestamp}
}
\value{
  integer value for the the new header level
}
\description{
  Increment header level
}
\details{
  Increases the header level by one, using the given header
  and clock time
}
\author{
  Dr Rainer M Krug
}

