% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_generator}
\alias{layout_generator}
\title{Generating logging function}
\usage{
layout_generator(msg_format = "{level} [{time}] {msg}",
  time_format = "\%Y-\%d-\%m \%H:\%M:\%S")
}
\arguments{
\item{msg_format}{\code{glue}-flavored layout of the message}

\item{time_format}{see \code{strptime} for details}

\item{level}{log level, eg \code{INFO}}

\item{msg}{character vector}
}
\value{
function taking \code{level} and \code{msg} arguments
}
\description{
Generating logging function
}
