% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_raw}
\alias{layout_raw}
\title{Format a log record by concatenating the log level, timestamp and message}
\usage{
layout_raw(level, msg)
}
\arguments{
\item{level}{log level, see \code{\link{log_levels}} for more details}

\item{msg}{string message}
}
\value{
character vector
}
\description{
Format a log record by concatenating the log level, timestamp and message
}
