% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_syslognet}
\alias{layout_syslognet}
\title{Format a log record for syslognet}
\usage{
layout_syslognet(
  level,
  msg,
  namespace = NA_character_,
  .logcall = sys.call(),
  .topcall = sys.call(-1),
  .topenv = parent.frame()
)
}
\arguments{
\item{level}{level of log message.}

\item{msg}{string of log message.}

\item{namespace}{namespace.}

\item{.logcall}{logcall; see \code{\link[logger]{layout_blank}}.}

\item{.topcall}{topcall; see \code{\link[logger]{layout_blank}}.}

\item{.topenv}{topenv; ; see \code{\link[logger]{layout_blank}}.}
}
\value{
A character vector with a severity attribute.
}
\description{
Format a log record for syslognet.
This function converts the logger log level to a
log severity level according to RFC 5424 "The Syslog Protocol".
}
