% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log}
\alias{log}
\alias{log_fatal}
\alias{log_error}
\alias{log_warn}
\alias{log_info}
\alias{log_debug}
\alias{log_trace}
\title{Log a message with given log level}
\usage{
log(level, msg)
}
\arguments{
\item{level}{log level from \code{log_levels}}

\item{msg}{character vector}
}
\description{
Log a message with given log level
}
\examples{
\dontrun{
log(INFO, 'hi there')
log_info('hi there')

## output omitted
log_debug('hi there')

## lower threshold and retry
log_threshold(TRACE)
log_debug('hi there')

## multiple lines
log_info('ok {1:3} + {1:3} = {2*(1:3)}')
log_layout(layout_json)

## note for the JSON output, glue is not automatically applied
log_info(glue::glue('ok {1:3} + {1:3} = {2*(1:3)}'))
}
}
