% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_chunk_time}
\alias{log_chunk_time}
\title{Automatically log execution time of knitr chunks}
\usage{
log_chunk_time(..., level = INFO, namespace = NA_character_)
}
\arguments{
\item{...}{passed to \code{\link[=log_level]{log_level()}}}

\item{level}{see \code{\link[=log_levels]{log_levels()}}}

\item{namespace}{x}
}
\description{
Calling this function in the first chunk of a document will instruct knitr
to automatically log the execution time of each chunk. If using
\code{\link[=formatter_glue]{formatter_glue()}} or \code{\link[=formatter_cli]{formatter_cli()}} then the \code{options} variable will be
available, providing the chunk options such as chunk label etc.
}
\examples{
# To be put in the first chunk of a document
log_chunk_time("chunk {options$label}")

}
