% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_config_setter}
\alias{log_config_setter}
\title{Base Logging Function}
\usage{
log_config_setter(fun_name, arg, namespace, index)
}
\arguments{
\item{fun_name}{string a full name of log function}

\item{arg}{see \code{\link[=log_levels]{log_levels()}}}

\item{namespace}{logger namespace}

\item{index}{index of the logger within the namespace}
}
\value{
If \code{arg} is \code{NULL}, will return the currently set value.
If \code{arg} is not \code{NULL}, will return the previously set value.
}
\description{
Base Logging Function
}
\keyword{internal}
