% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_failure}
\alias{log_failure}
\title{Logs the error message to console before failing}
\usage{
log_failure(expression)
}
\arguments{
\item{expression}{call}
}
\description{
Logs the error message to console before failing
}
\examples{
log_failure("foobar")
try(log_failure(foobar))
}
