% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_formatter}
\alias{log_formatter}
\title{Get or set log message formatter}
\usage{
log_formatter(formatter = NULL, namespace = "global", index = 1)
}
\arguments{
\item{formatter}{function defining how R objects are converted into
a single string, eg \code{\link[=formatter_paste]{formatter_paste()}}, \code{\link[=formatter_sprintf]{formatter_sprintf()}},
\code{\link[=formatter_glue]{formatter_glue()}}, \code{\link[=formatter_glue_or_sprintf]{formatter_glue_or_sprintf()}},
\code{\link[=formatter_logging]{formatter_logging()}}, default NULL}

\item{namespace}{logger namespace}

\item{index}{index of the logger within the namespace}
}
\description{
Get or set log message formatter
}
\seealso{
Other log configutation functions:
\code{\link[=log_appender]{log_appender()}},
\code{\link[=log_layout]{log_layout()}},
\code{\link[=log_threshold]{log_threshold()}}
}
\concept{log configutation functions}
