% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_indices}
\alias{log_indices}
\title{Returns number of currently active indices}
\usage{
log_indices(namespace = "global")
}
\arguments{
\item{namespace}{override the default / auto-picked namespace with
a custom string}
}
\value{
number of indices
}
\description{
Returns number of currently active indices
}
