% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\name{log_levels}
\alias{log_levels}
\alias{OFF}
\alias{FATAL}
\alias{ERROR}
\alias{WARN}
\alias{SUCCESS}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{Log levels}
\usage{
OFF

FATAL

ERROR

WARN

SUCCESS

INFO

DEBUG

TRACE
}
\description{
The standard Apache logj4 log levels plus a custom level for
\code{SUCCESS}. For the full list of these log levels and suggested
usage, check the below Details.
}
\details{
List of supported log levels:
\itemize{
\item \code{OFF} No events will be logged
\item \code{FATAL} Severe error that will prevent the application from continuing
\item \code{ERROR} An error in the application, possibly recoverable
\item \code{WARN} An event that might possible lead to an error
\item \code{SUCCESS} An explicit success event above the INFO level that you want to log
\item \code{INFO} An event for informational purposes
\item \code{DEBUG} A general debugging event
\item \code{TRACE} A fine-grained debug message, typically capturing the flow through the application.
}
}
\references{
\url{https://logging.apache.org/log4j/2.x/javadoc/log4j-api/org/apache/logging/log4j/Level.html},
\url{https://logging.apache.org/log4j/2.x/manual/customloglevels.html}
}
