% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{log_messages}
\alias{log_messages}
\title{Injects a logger call to standard messages}
\usage{
log_messages()
}
\description{
This function uses \code{\link[=trace]{trace()}} to add a \code{\link[=log_info]{log_info()}} function call when
\code{message} is called to log the informative messages with the
\code{logger} layout and appender.
}
\examples{
\dontrun{
log_messages()
message("hi there")
}
}
