% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{log_shiny_input_changes}
\alias{log_shiny_input_changes}
\title{Auto logging input changes in Shiny app}
\usage{
log_shiny_input_changes(
  input,
  level = INFO,
  namespace = NA_character_,
  excluded_inputs = character(),
  allow = FALSE,
  change_message = "Shiny input change detected on {name}: {old} -> {new}",
  initialize_message = "Default Shiny inputs initialized:"
)
}
\arguments{
\item{input}{passed from Shiny's \code{server}}

\item{level}{log level}

\item{namespace}{the name of the namespace}

\item{excluded_inputs}{character vector of input names to exclude from logging}

\item{allow}{allow to be used outside Shiny app}

\item{change_message}{custom message to be displayed during the input change - supports `glue` syntax where `{name}`
is the name of the input, and `{old}` and `{new}` are input values before the change and after the change}

\item{initialize_message}{custom message to be displayed during input initialization (this is followed by initialized
input values being printed in JSON format)}
}
\description{
This is to be called in the \code{server} section of the Shiny app.
}
\examples{
\dontrun{
library(shiny)

ui <- bootstrapPage(
    numericInput('mean', 'mean', 0),
    numericInput('sd', 'sd', 1),
    textInput('title', 'title', 'title'),
    textInput('foo', 'This is not used at all, still gets logged', 'foo'),
    passwordInput('password', 'Password not to be logged', 'secret'),
    plotOutput('plot')
)

server <- function(input, output) {

    logger::log_shiny_input_changes(input, excluded_inputs = 'password')

    output$plot <- renderPlot({
        hist(rnorm(1e3, input$mean, input$sd), main = input$title)
    })

}

shinyApp(ui = ui, server = server)
}
}
