% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_tictoc}
\alias{log_tictoc}
\title{Tic-toc logging}
\usage{
log_tictoc(..., level = INFO, namespace = NA_character_)
}
\arguments{
\item{...}{passed to \code{log_level}}

\item{level}{see \code{\link[=log_levels]{log_levels()}}}

\item{namespace}{x}
}
\description{
Tic-toc logging
}
\examples{
\dontrun{
log_tictoc('warming up')
Sys.sleep(0.1)
log_tictoc('running')
Sys.sleep(0.1)
log_tictoc('running')
Sys.sleep(runif(1))
log_tictoc('and running')
}
}
\author{
Thanks to Neal Fultz for the idea and original implementation!
}
