% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_with_separator}
\alias{log_with_separator}
\title{Logs a message in a very visible way}
\usage{
log_with_separator(..., level = INFO, namespace = NA_character_,
  separator = "=")
}
\arguments{
\item{message}{string}
}
\description{
Logs a message in a very visible way
}
\examples{
log_with_separator('An important message')
log_with_separator('This message is worth a {1e3} words')
log_with_separator('A very important message with a bunch of extra words that will eventually wrap into a multi-line message for our quite nice demo :wow:')
log_with_separator('Boo!', level = FATAL)
}
