\name{logBegin}
\alias{logBegin}
\title{logBegin}
\usage{
  logBegin(...)
}
\arguments{
  \item{...}{values which will be concatenated using
  paste(..., collapse=" ") to the message to be logged.}
}
\value{
  invisible returns new header level
}
\description{
  log as org mode compatible format
}
\details{
  Logging in org mode compatible format
}
\author{
  Rainer M Krug
}

