% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger-package.R
\docType{package}
\name{logger-package}
\alias{logger-package}
\title{logger: A Lightweight, Modern and Flexible Logging Utility}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Inspired by the the 'futile.logger' R package and 'logging' Python module, this utility provides a flexible and extensible way of formatting and delivering log messages with low overhead.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://daroczig.github.io/logger/}
  \item Report bugs at \url{https://github.com/daroczig/logger/issues}
}

}
\author{
\strong{Maintainer}: Gergely Daróczi \email{daroczig@rapporter.net} (\href{https://orcid.org/0000-0003-3149-8537}{ORCID})

Authors:
\itemize{
  \item Gergely Daróczi \email{daroczig@rapporter.net} (\href{https://orcid.org/0000-0003-3149-8537}{ORCID})
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
}

Other contributors:
\itemize{
  \item Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID}) [contributor]
  \item Spare Cores [funder]
  \item System1 [funder]
}

}
\keyword{internal}
