% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.getLevel}
\alias{logger.getLevel}
\title{Get configured logging level.}
\usage{
logger.getLevel()
}
\value{
A string giving the lowest message level that will be reported, one of
"ALL", "DEBUG", "INFO", "WARN", "ERROR", "SEVERE", or "OFF".
}
\description{
This will return the current level configured of the logging messages.
}
\details{
Note that \code{logger.setLevel()} invisibly returns current level, so
\code{logger.getLevel()} is not required to restore the level after a
temporary change.
}
\examples{
\dontrun{
logger.getLevel()
}
}
\author{
Rob Kooper
}
