% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Generate logging utility}
\usage{
logger(threshold, formatter, layout, appender)
}
\arguments{
\item{threshold}{omit log messages below this \code{log_levels}}

\item{formatter}{function pre-processing the message of the log record}

\item{layout}{function rendering the layout of the actual log record}

\item{appender}{function writing the log record}
}
\value{
function taking \code{level} and \code{msg} arguments
}
\description{
Generate logging utility
}
