% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.setLevel}
\alias{logger.setLevel}
\title{Configure logging level.}
\usage{
logger.setLevel(level)
}
\arguments{
\item{level}{the level of the message. One of "ALL", "DEBUG", "INFO", "WARN",
"ERROR", "SEVERE", or "OFF".}
}
\value{
When logger level is set, the previous level is returned invisibly.
This can be passed to \code{logger.setLevel()} to restore the previous level.
}
\description{
This will configure the logger level. This allows to turn DEBUG, INFO,
WARN, ERROR, and SEVERE messages on and off.
}
\details{
Note that this controls \emph{printing} of messages and does not change other behavior.
In particular, suppressing SEVERE by setting the level to "OFF" does not prevent
logger.severe() from signaling an error (and terminating the program if
\code{logger.setQuitOnSevere(TRUE)}).
}
\examples{
\dontrun{
logger.setLevel('DEBUG')

# Temporarily turn logger off
old_logger_level <- logger.setLevel("OFF")
  # code here
logger.setLevel(old_logger_level)
}
}
\author{
Rob Kooper
}
