% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.setQuitOnSevere}
\alias{logger.setQuitOnSevere}
\title{Configure whether severe should quit.}
\usage{
logger.setQuitOnSevere(severeQuits)
}
\arguments{
\item{severeQuits}{should R quit on a severe error.}
}
\value{
invisibly, the previous value of \code{severeQuits}.
This can be used to restore settings after a temporary change.
}
\description{
The default is for a non-interactive session to quit.
Setting this to false is especially useful for running tests.
}
\examples{
\dontrun{
logger.setQuitOnSevere(FALSE)
}
}
\author{
Rob Kooper
}
