% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.setUseConsole}
\alias{logger.setUseConsole}
\title{Configure logging to console.}
\usage{
logger.setUseConsole(console, stderr = TRUE)
}
\arguments{
\item{console}{set to true to print logging to console.}

\item{stderr}{set to true (default) to log to stderr instead of stdout}
}
\value{
Invisibly, a list of the previously set values of \code{console}
and \code{stderr}. This can be used to restore the previous settings after a
temporary change.
}
\description{
Should the logging to be printed to the console or not.
}
\examples{
\dontrun{
logger.setUseConsole(TRUE)
}
}
\author{
Rob Kooper
}
