\name{logList}
\alias{logList}
\title{logList}
\usage{
  logList(...)
}
\arguments{
  \item{...}{values which will be concatenated using
  paste(..., collapse=" ") to the message to be logged.}
}
\description{
  log an entry without changing the header level
}
\details{
  Log a message in the actual logging level.
}
\author{
  Rainer M Krug
}

