% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkLeaflet.R
\name{mkLeaflet}
\alias{mkLeaflet}
\title{\code{mkLeaflet} Make interactive Leaflet map of tracks and sightings}
\usage{
mkLeaflet(data.list = list(midnat, fram), birds = F)
}
\arguments{
\item{data.list}{List of sublists, each of which contains data frames created using \code{readLogger}}

\item{birds}{Logical, indicating if map should present birds (TRUE) or whales (FALSE).
Default is to present whales. Both cannot (currently) be shown on the same map.}
}
\value{
Saves a Leaflet object as an html file called \code{map.html} in the working directory
}
\description{
\code{mkLeaflet} Make interactive Leaflet map of tracks and sightings
}
\details{
Draws an interactive map in a web browser, where the background is a WMS data layer.
NOTE! Computer must currently be connected to the internet in order to render the map background!
}
\seealso{
\code{\link{readLoggerTable}} to read single table from sightings database,
  \code{\link{readLogger}} to read entire sightings database,
  \code{\link{correctEffort}} tto correct inconsistencies in effort table,
  \code{\link{filterGPS}} to flag bad GPS positions,
  \code{\link{fillGPS}} To fill in missing GPS positions after filtering

Other Logger sightings database functions: 
\code{\link{correctEffort}()},
\code{\link{fillGPS}()},
\code{\link{filterGPS}()},
\code{\link{readLoggerTable}()},
\code{\link{readLogger}()}
}
\author{
Martin Biuw
}
\concept{Logger sightings database functions}
