% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print2string.R
\name{print2string}
\alias{print2string}
\title{Convert the printed output of an object to a string. This is
particularly useful for passing complex objects (e.g. matrices,
data.frames) to simple printing functions like
\code{\link[=logger.info]{logger.info()}} or \code{\link[base:message]{base::message()}}.}
\usage{
print2string(x, ...)
}
\arguments{
\item{x}{Object to print}

\item{...}{Additional arguments to \code{\link[base:print]{base::print()}}.}
}
\value{
Output of \code{print(x, ...)}, captured as string
}
\description{
Note that for this to work properly in the \code{PEcAn.logger}
functions, you should always add the \code{wrap = FALSE} argument, and
probably add a newline (\code{"\\n"}) before the output of this function.
}
\examples{
logger.info("First few rows of Iris:\n", print2string(iris[1:10, -5]), wrap = FALSE)
df <- data.frame(test = c("download", "process", "plot"), status = c(TRUE, TRUE, FALSE))
logger.debug("Current status:\n", print2string(df, row.names = FALSE), wrap = FALSE)
}
\author{
Alexey Shiklomanov
}
