% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerfiles.R
\name{read_loggerfiles}
\alias{read_loggerfiles}
\title{Read and combine logger csv files}
\usage{
read_loggerfiles(loggerinfo, folder = "data/raw_csv", check_SN = TRUE)
}
\arguments{
\item{loggerinfo}{The name of the `dataframe`containing the logger metadata}

\item{folder}{The path to the folder containing the logger `.csv` files. The default is `data/raw_csv`}

\item{check_SN}{logical: Should the serial number (SN) provided in the metadata dataframe be checked against the SN stored in the logger `.csv` files? The default is `TRUE`.}
}
\value{
A `dataframe` in "long" format containing all the logger data. Additional information, as SN, filename, logger type, etc, is added
}
\description{
Read and combine logger csv files
}
\examples{
\dontrun{
dat_logger <- read_loggerfiles(loggerinfo)
}

}
