% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerinfo.R
\name{read_loggerinfo}
\alias{read_loggerinfo}
\title{Read logger metadata}
\usage{
read_loggerinfo(path, tab = "retrieved")
}
\arguments{
\item{path}{The path to the Excel file (`.xlsx`) containing the metadata for the logger deployments}

\item{tab}{The tab in the Excel file containing the metadata for the logger deployments}
}
\value{
a `data.frame` with the logger metadata
}
\description{
This function is a wrapper for `readxl::read_excel` and imports the Excel table containing the logger metadata.
}
\examples{
\dontrun{
loggerinfo <- read_loggerinfo(here::here("data/logger_info.xlsx"))
}

}
