% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLogger.R
\name{readLogger}
\alias{readLogger}
\title{\code{readLogger} Read all tables from Logger Access sightings database}
\usage{
readLogger(
  DB = "Midnatsol_20191122",
  whales = T,
  birds = T,
  spatial = F,
  interval = 60
)
}
\arguments{
\item{DB}{Name of database as defined in ODBC data sources}

\item{whales}{Logical, indicating if whale sightings table should be read}

\item{birds}{Logical, indicating if bird sightings tables should be read}

\item{spatial}{Logical indicating if a table containing GPS data
should be returned as a data frame (default) or as a SpatialPointsDataFrame}

\item{interval}{Interval setting the downscaling of GPS data.
Read in every \code{interval} line from data table.}
}
\value{
Returns a list of data frames corresponding to desired tables in the database.
}
\description{
\code{readLogger} Read all tables from Logger Access sightings database
}
\details{
The database must first be set up as a DSN data source.
  Note! On Windows, only 32-bit data sources can be used,
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
}
\seealso{
\code{\link{readLoggerTable}} to read single table from sightings database,
  \code{\link{correctEffort}} to correct inconsistencies in effort table,
  \code{\link{filterGPS}} to flag bad GPS positions,
  \code{\link{fillGPS}} to fill in missing GPS positions after filtering
  \code{\link{mkLeaflet}} to make interactive Leaflet map of sightings.
  \code{\link{convert2spatial}} to convert selected parts of sightings data list to shapefiles.

Other Logger sightings database functions: 
\code{\link{convert2spatial}()},
\code{\link{correctEffort}()},
\code{\link{fillGPS}()},
\code{\link{filterGPS}()},
\code{\link{mkLeaflet}()},
\code{\link{readLoggerTable}()}
}
\author{
Martin Biuw
}
\concept{Logger sightings database functions}
