% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLogger.R
\name{readLoggerTable}
\alias{readLoggerTable}
\title{\code{readLoggerTable} Read table from Logger Access database}
\usage{
readLoggerTable(
  theDB = "Midnatsol_20191122",
  theTable = "GpsData",
  to.spatial = F,
  int = 1
)
}
\arguments{
\item{theDB}{Name of database as defined in ODBC data sources}

\item{theTable}{Name of the table to read}

\item{to.spatial}{Logical indicating if a table containing GPS data
should be returned as a data frame (default) or as a SpatialPointsDataFrame}

\item{int}{Interval setting the downscaling of GPS data.
Read in every \code{int} line from data table.}
}
\value{
Returns a data frame corresponding to the desired table.
}
\description{
\code{readLoggerTable} Read table from Logger Access database
}
\details{
The database must first be set up as a DSN data source.
  Note! On Windows, only 32-bit data sources can be used,
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
}
\seealso{
\code{\link{readLogger}} to read entire sightings database,
  \code{\link{correctEffort}} to correct inconsistencies in effort table,
  \code{\link{filterGPS}} to flag bad GPS positions,
  \code{\link{mkLeaflet}} to make interactive Leaflet map of sightings.

Other Logger sightings database functions: 
\code{\link{correctEffort}()},
\code{\link{fillGPS}()},
\code{\link{filterGPS}()},
\code{\link{mkLeaflet}()},
\code{\link{readLogger}()}
}
\author{
Martin Biuw
}
\concept{Logger sightings database functions}
