\name{setLogFileName}
\alias{setLogFileName}
\title{Set log file name}
\usage{
  setLogFileName(newLogFileName = NULL)
}
\arguments{
  \item{newLogFileName}{the new log file name to be used
  default value: 10}
}
\value{
  the old logging threshold used
}
\description{
  Set log file name If the name is null (default), logging
  to file will be disabled.
}
\author{
  Dr Rainer M Krug
}

