\name{setLoggingThreshold}
\alias{setLoggingThreshold}
\title{Set logging threshold}
\usage{
  setLoggingThreshold(newLoggingThreshold = 10)
}
\arguments{
  \item{newLoggingThreshold}{the new logging threshold to
  be used.  default value: 10}
}
\value{
  the old logging threshold used
}
\description{
  Set logging threshold to new value
}
\author{
  Dr Rainer M Krug
}

