\name{setLogToConsole}
\alias{setLogToConsole}
\title{Set logToConsole}
\usage{
  setLogToConsole(newValue = NULL)
}
\arguments{
  \item{newValue}{logical, TRUE if logs should be written
  to console, FALSE if not default value: FALSE}
}
\value{
  the old value of logToConsole
}
\description{
  Set logToConsole
}
\author{
  Dr Rainer M Krug
}

