\name{setTimeFormat}
\alias{setTimeFormat}
\title{Set time format}
\usage{
  setTimeFormat(newFormat)
}
\arguments{
  \item{newFormat}{new timeformat to be used. See ?strptime
  for details of the format}
}
\value{
  the old format string used
}
\description{
  Set time format which will be used to formate the
  timestamps
}
\author{
  Dr Rainer M Krug
}

