% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logifnot.R
\name{severeifnot}
\alias{severeifnot}
\alias{errorifnot}
\alias{warnifnot}
\alias{infoifnot}
\alias{debugifnot}
\title{Logger message if conditions are not met}
\usage{
severeifnot(msg, ...)

errorifnot(msg, ...)

warnifnot(msg, ...)

infoifnot(msg, ...)

debugifnot(msg, ...)
}
\arguments{
\item{msg}{Logger message to write, as a single character string.}

\item{...}{Conditions to evaluate}
}
\value{
Invisibly, \code{TRUE} if conditions are met, \code{FALSE} otherwise
}
\description{
Similar to \link[base:stopifnot]{base::stopifnot}, but allows you to use a custom message and
logger level. If all conditions are \code{TRUE}, silently exit.
}
\details{
Conditions can be vectorized, or can return non-logical values.The
underlying function automatically applies \code{isTRUE(all(.))} to the
conditions.
}
\examples{
a <- 1:5
b <- list(6, 7, 8)
debugifnot("By the way, something is not a list.", is.list(a), is.list(b))
infoifnot("Something is not a list.", is.list(a), is.list(b))
warnifnot("I would prefer it if you used lists.", is.list(a), is.list(b))
errorifnot("You should definitely use lists.", is.list(a), is.list(b))
try({
  severeifnot(
    "I cannot deal with the fact that something is not a list.",
    is.list(a),
    is.list(b)
  )
})
}
