% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{skip_formatter}
\alias{skip_formatter}
\title{Adds the skip_formatter attribute to an object so that logger will skip calling the formatter function on the object(s) to be logged}
\usage{
skip_formatter(message, ...)
}
\arguments{
\item{message}{character vector directly passed to the appender function in \code{\link{logger}}}

\item{...}{should be never set}
}
\value{
character vector  with \code{skip_formatter} attribute set to \code{TRUE}
}
\description{
Adds the skip_formatter attribute to an object so that logger will skip calling the formatter function on the object(s) to be logged
}
