% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_daily.R
\name{summarise_daily}
\alias{summarise_daily}
\title{Summarize daily mean, min, and max values}
\usage{
summarise_daily(loggerdata, by = "filename", parameter = "temperature")
}
\arguments{
\item{loggerdata}{A dataframe containing the `loggerdada`}

\item{by}{A character string of columns in `loggerdata` by which the data should be grouped for the summary, additional to `filename`.}

\item{parameter}{The parameter that should be summarised. Can be `temperature` or `light_intensity`.}
}
\value{
A dataframe containing the columns specified in `by` and the mean, min, and max values of `parameter`.
}
\description{
This function summarizes each logger run for the daily mean, min, and max values. This is supported fro `temperature` and `light_intensity`.
}
\examples{
\dontrun{
loggerdataS <- summarise_daily(loggerdata)
}



}
