% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{top_env_name}
\alias{top_env_name}
\title{Returns the name of the top level environment from which the logger was called}
\usage{
top_env_name(.topenv = parent.frame())
}
\arguments{
\item{.topenv}{call environment}
}
\value{
string
}
\description{
Returns the name of the top level environment from which the logger was called
}
\keyword{internal}
