% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{validate_log_level}
\alias{validate_log_level}
\title{Assure valid log level}
\usage{
validate_log_level(level)
}
\arguments{
\item{level}{\code{\link[=log_levels]{log_levels()}} object or string representation}
}
\value{
\code{\link[=log_levels]{log_levels()}} object
}
\description{
Assure valid log level
}
\keyword{internal}
