% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{warn_if_globalCallingHandlers_is_not_available}
\alias{warn_if_globalCallingHandlers_is_not_available}
\title{Warn to update R to 4+}
\usage{
warn_if_globalCallingHandlers_is_not_available()
}
\description{
Warn to update R to 4+
}
\keyword{internal}
