% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_async}
\alias{appender_async}
\title{Delays executing the actual appender function to the future in a background process to avoid blocking the main R session}
\usage{
appender_async(appender, batch = 1)
}
\arguments{
\item{appender}{a  \code{\link{log_appender}} function with a \code{generator} attribute (TODO note not required, all fn will be passed if not)}

\item{batch}{number of records to process from the queue at once}
}
\value{
function taking \code{lines} argument
}
\description{
Delays executing the actual appender function to the future in a background process to avoid blocking the main R session
}
\note{
This functionality depends on the \pkg{txtq} and \pkg{callr} packages.
}
\examples{
\dontrun{
appender_file_slow <- function(file) {
  function(lines) {
    Sys.sleep(5)
    cat(lines, sep = '\\n', file = file, append = TRUE)
  }
}

t <- tempfile()
log_appender(appender_async(appender_file_slow(file = t)))
log_info('Was this slow?')
}
}
