% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_console}
\alias{appender_console}
\title{Append log messages to stdout}
\usage{
appender_console(lines)
}
\arguments{
\item{lines}{character vector}
}
\description{
Append log messages to stdout
}
