% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_console}
\alias{appender_console}
\title{Append log record to stdout}
\usage{
appender_console(lines)
}
\arguments{
\item{lines}{character vector}
}
\description{
Append log record to stdout
}
\seealso{
This is a \code{\link{log_appender}}, for alternatives, see eg \code{\link{appender_file}} or \code{\link{appender_tee}}
}
