% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_file}
\alias{appender_file}
\title{Append log messages to a file}
\usage{
appender_file(file)
}
\arguments{
\item{file}{path}
}
\value{
function taking \code{lines} argument
}
\description{
Append log messages to a file
}
