% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_kinesis}
\alias{appender_kinesis}
\title{Send log messages to a Amazon Kinesis stream}
\usage{
appender_kinesis(stream)
}
\arguments{
\item{stream}{name of the Kinesis stream}
}
\value{
function taking \code{lines} and optional \code{partition_key} argument
}
\description{
Send log messages to a Amazon Kinesis stream
}
\note{
This functionality depends on the \pkg{botor} package.
}
\seealso{
This is generator function for \code{\link{log_appender}}, for alternatives, see eg \code{\link{appender_console}}, \code{\link{appender_file}}, \code{\link{appender_tee}}, \code{\link{appender_pushbullet}}, \code{\link{appender_syslog}} and \code{\link{appender_async}} for evaluate any \code{\link{log_appender}} function in a background process.
}
