% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_kinesis}
\alias{appender_kinesis}
\title{Send log messages to a Amazon Kinesis stream}
\usage{
appender_kinesis(stream)
}
\arguments{
\item{stream}{name of the Kinesis stream}
}
\value{
function taking \code{lines} and optional \code{partition_key}
argument
}
\description{
Send log messages to a Amazon Kinesis stream
}
\note{
This functionality depends on the \pkg{botor} package.
}
\seealso{
Other log_appenders: 
\code{\link{appender_async}()},
\code{\link{appender_console}()},
\code{\link{appender_file}()},
\code{\link{appender_ntfy}()},
\code{\link{appender_pushbullet}()},
\code{\link{appender_slack}()},
\code{\link{appender_stdout}()},
\code{\link{appender_syslog}()},
\code{\link{appender_tee}()},
\code{\link{appender_telegram}()}
}
\concept{log_appenders}
