% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_stdout}
\alias{appender_stdout}
\title{Append log record to stdout}
\usage{
appender_stdout(lines)
}
\arguments{
\item{lines}{character vector}
}
\description{
Append log record to stdout
}
\seealso{
This is a \code{\link{log_appender}}, for alternatives, see eg \code{\link{appender_console}}, \code{\link{appender_file}}, \code{\link{appender_tee}}, \code{\link{appender_slack}}, \code{\link{appender_pushbullet}}
}
