% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appenders.R
\name{appender_telegram}
\alias{appender_telegram}
\title{Send log messages to a Telegram chat}
\usage{
appender_telegram(chat_id = Sys.getenv("TELEGRAM_CHAT_ID"),
  bot_token = Sys.getenv("TELEGRAM_BOT_TOKEN"), parse_mode = NULL)
}
\arguments{
\item{chat_id}{Unique identifier for the target chat or username of the target channel (in the format @channelusername)}

\item{bot_token}{Telegram Authorization token}

\item{parse_mode}{Message parse mode. Allowed values: Markdown or HTML}
}
\value{
function taking \code{lines} argument
}
\description{
Send log messages to a Telegram chat
}
\note{
This functionality depends on the \pkg{telegram} package.
}
\seealso{
This is generator function for \code{\link{log_appender}}, for alternatives, see eg \code{\link{appender_console}}, \code{\link{appender_file}}, \code{\link{appender_tee}}, \code{\link{appender_pushbullet}}, \code{\link{appender_syslog}}, \code{\link{appender_kinesis}} and \code{\link{appender_async}} for evaluate any \code{\link{log_appender}} function in a background process.
}
