% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catch_base_log}
\alias{catch_base_log}
\title{Catch the log header}
\usage{
catch_base_log(level, namespace)
}
\arguments{
\item{level}{see \code{\link{log_levels}}}

\item{namespace}{string}
}
\value{
string
}
\description{
Catch the log header
}
\keyword{internal}
