% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse_to_one_line}
\alias{deparse_to_one_line}
\title{Deparse and join all lines into a single line}
\usage{
deparse_to_one_line(x)
}
\arguments{
\item{x}{object to \code{deparse}}
}
\value{
string
}
\description{
We do not recommend using this function in new code; it only exported
for historical reasons.
}
\keyword{internal}
