% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fail_on_missing_package}
\alias{fail_on_missing_package}
\title{Check if R package can be loaded and fails loudly otherwise}
\usage{
fail_on_missing_package(pkg)
}
\arguments{
\item{pkg}{string}
}
\description{
Check if R package can be loaded and fails loudly otherwise
}
\keyword{internal}
