% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_cli}
\alias{formatter_cli}
\title{Apply \code{\link[cli:cli_text]{cli::cli_text()}} to format string with cli syntax}
\usage{
formatter_cli(
  ...,
  .logcall = sys.call(),
  .topcall = sys.call(-1),
  .topenv = parent.frame()
)
}
\arguments{
\item{...}{passed to \code{\link[cli:cli_text]{cli::cli_text()}} for the text interpolation}

\item{.logcall}{the logging call being evaluated (useful in
formatters and layouts when you want to have access to the raw,
unevaluated R expression)}

\item{.topcall}{R expression from which the logging function was
called (useful in formatters and layouts to extract the calling
function's name or arguments)}

\item{.topenv}{original frame of the \code{.topcall} calling function
where the formatter function will be evaluated and that is used
to look up the \code{namespace} as well via \code{logger:::top_env_name}}
}
\value{
character vector
}
\description{
Apply \code{\link[cli:cli_text]{cli::cli_text()}} to format string with cli syntax
}
\seealso{
Other log_formatters: 
\code{\link{formatter_glue}()},
\code{\link{formatter_glue_or_sprintf}()},
\code{\link{formatter_glue_safe}()},
\code{\link{formatter_json}()},
\code{\link{formatter_logging}()},
\code{\link{formatter_pander}()},
\code{\link{formatter_paste}()},
\code{\link{formatter_sprintf}()}
}
\concept{log_formatters}
