% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_glue}
\alias{formatter_glue}
\title{Apply \code{glue}}
\usage{
formatter_glue(...)
}
\arguments{
\item{...}{passed to \code{glue} for the text interpolation}
}
\value{
character vector
}
\description{
Apply \code{glue}
}
