% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_glue}
\alias{formatter_glue}
\title{Concatenate strings and apply glue formatter}
\usage{
formatter_glue(...)
}
\arguments{
\item{...}{first passed to \code{paste} to generate a character vector of all passed objects, then passed to \code{glue}}
}
\value{
character vector
}
\description{
Concatenate strings and apply glue formatter
}
