% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_glue_or_sprinf}
\alias{formatter_glue_or_sprinf}
\title{Apply \code{glue} and \code{sprintf}}
\usage{
formatter_glue_or_sprinf(msg, ...)
}
\arguments{
\item{...}{passed to \code{glue} for the text interpolation}
}
\value{
character vector
}
\description{
The best of both words: using both formatter functions in your log messages, which can be useful eg if you are migrating from \code{sprintf} formatted log messages to \code{glue} or similar.
}
\details{
Note that this function tries to be smart when passing arguments to \code{glue} and \code{sprintf}, but might fail with some edge cases, and returns an unformatted string.
}
\examples{
\dontrun{
formatter_glue_or_sprinf("{a} + {b} = \%s", a = 2, b = 3, 5)
formatter_glue_or_sprinf("{pi} * {2} = \%s", pi*2)
formatter_glue_or_sprinf("{pi} * {2} = {pi*2}")

formatter_glue_or_sprinf("Hi ", "{c('foo', 'bar')}, did you know that 2*4={2*4}")
formatter_glue_or_sprinf("Hi {c('foo', 'bar')}, did you know that 2*4={2*4}")
formatter_glue_or_sprinf("Hi {c('foo', 'bar')}, did you know that 2*4=\%s", 2*4)
formatter_glue_or_sprinf("Hi \%s, did you know that 2*4={2*4}", c('foo', 'bar'))
formatter_glue_or_sprinf("Hi \%s, did you know that 2*4=\%s", c('foo', 'bar'), 2*4)
}
}
