% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_glue_or_sprinf}
\alias{formatter_glue_or_sprinf}
\title{Apply \code{glue} and \code{sprintf}}
\usage{
formatter_glue_or_sprinf(msg, ...)
}
\arguments{
\item{...}{passed to \code{glue} for the text interpolation}
}
\value{
character vector
}
\description{
The best of both words: using both formatter functions in your log messages, which can be useful eg if you are migrating from \code{sprintf} formatted log messages to \code{glue} or similar.
}
\details{
Note that this function tries to be smart when passing arguments to \code{glue} and \code{sprintf}, but might fail with some edge cases, and returns an unformatted string.
}
