% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_paste}
\alias{formatter_paste}
\title{Concatenate strings via \code{paste}}
\usage{
formatter_paste(...)
}
\arguments{
\item{...}{R objects that can be converted to string}
}
\value{
string
}
\description{
Concatenate strings via \code{paste}
}
