% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_sprintf}
\alias{formatter_sprintf}
\title{Apply \code{sprintf}}
\usage{
formatter_sprintf(fmt, ...)
}
\arguments{
\item{...}{passed to \code{sprintf}}
}
\value{
character vector
}
\description{
Apply \code{sprintf}
}
