% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_sprintf}
\alias{formatter_sprintf}
\title{Apply \code{sprintf} to convert R objects into a character vector}
\usage{
formatter_sprintf(fmt, ..., .envir = parent.frame(1))
}
\arguments{
\item{fmt}{passed to \code{sprintf}}

\item{...}{passed to \code{sprintf}}
}
\value{
character vector
}
\description{
Apply \code{sprintf} to convert R objects into a character vector
}
\seealso{
This is a \code{\link{log_formatter}}, for alternatives, see \code{\link{formatter_paste}}, \code{\link{formatter_glue}}, \code{\link{formatter_glue_or_sprintf}}
}
