% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{formatter_sprintf}
\alias{formatter_sprintf}
\title{Apply \code{sprintf} to convert R objects into a character vector}
\usage{
formatter_sprintf(fmt, ..., .logcall = sys.call(),
  .topcall = sys.call(-1), .topenv = parent.frame())
}
\arguments{
\item{fmt}{passed to \code{sprintf}}

\item{...}{passed to \code{sprintf}}

\item{.logcall}{the logging call being evaluated (useful in formatters and layouts when you want to have access to the raw, unevaluated R expression)}

\item{.topcall}{R expression from which the logging function was called (useful in formatters and layouts to extract the calling function's name or arguments)}

\item{.topenv}{original frame of the \code{.topcall} calling function where the formatter function will be evaluated and that is used to look up the \code{namespace} as well via \code{\link{logger::top_env_name}}}
}
\value{
character vector
}
\description{
Apply \code{sprintf} to convert R objects into a character vector
}
\seealso{
This is a \code{\link{log_formatter}}, for alternatives, see \code{\link{formatter_paste}}, \code{\link{formatter_glue}}, \code{\link{formatter_glue_or_sprintf}}
}
