% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{get_logger_meta_variables}
\alias{get_logger_meta_variables}
\title{Collect useful information about the logging environment to be used in log messages}
\usage{
get_logger_meta_variables(log_level = NULL, .call = sys.call(-1),
  .envir = parent.frame())
}
\arguments{
\item{log_level}{log level as per \code{\link{log_levels}}}

\item{.call}{R expression from which the logging function was called (optionally used by the logging layout)}

\item{.envir}{original frame of the \code{.call} calling function where the formatter function is to be evaluated and that is used to look up the \code{namespace} as well via \code{\link{topenv}}}
}
\value{
list
}
\description{
Available variables to be used in the log formatter functions, eg in \code{\link{layout_glue_generator}}:
\itemize{
  \item levelr: log level as an R object, eg \code{\link{INFO}}
  \item level: log level as a string, eg \code{\link{INFO}}
  \item time: current time as \code{POSIXct}
  \item node: name by which the machine is known on the network as reported by \code{Sys.info}
  \item arch: machine type, typically the CPU architecture
  \item os_name: Operating System's name
  \item os_release: Operating System's release
  \item os_version: Operating System's version
  \item user: name of the real user id as reported by \code{Sys.info}
  \item pid: the process identification number of the R session
  \item node: name by which the machine is known on the network as reported by \code{Sys.info}
  \item namespace: R package (if any) calling the logging function
  \item call: parent call (if any) calling the logging function
  \item fn: function's (if any) name calling the logging function
}
}
\seealso{
\code{\link{layout_glue_generator}}
}
