% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{get_logger_meta_variables}
\alias{get_logger_meta_variables}
\title{Collect useful information about the logging environment to be used in log messages}
\usage{
get_logger_meta_variables(log_level)
}
\arguments{
\item{log_level}{log level as per \code{\link{log_levels}}}
}
\value{
list
}
\description{
Collect useful information about the logging environment to be used in log messages
}
\seealso{
layout_glue_generator
}
\keyword{internal}
