% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_generator}
\alias{layout_generator}
\title{Generate logging function}
\usage{
layout_generator(msg_format = "{level} [{time}] {msg}",
  time_format = "\%Y-\%d-\%m \%H:\%M:\%S")
}
\arguments{
\item{msg_format}{\code{glue}-flavored layout of the message}

\item{time_format}{see \code{strptime} for details}

\item{level}{log level, eg \code{INFO}}

\item{msg}{character vector}
}
\value{
function taking \code{level} and \code{msg} arguments - keeping the original call creating the generator in the \code{generator} attribute that is returned when calling \code{log_layout()} for the currently used layout
}
\description{
Available variables to be used in the \code{msg_format} provided by \code{logger:::get_logger_meta_variables}:
\itemize{
  \item level: log level, eg INFO
  \item time: current time formatted as \code{time_format}
  \item namespace: R package (if any) calling the logging function
  \item call: parent call (if any) calling the logging function
  \item fn: function's (if any) name calling the logging function
  \item user: name of the real user id as reported by \code{Sys.info}
  \item pid: the process identification number of the R session
  \item node: name by which the machine is known on the network as reported by \code{Sys.info}
}
}
\examples{
\dontrun{
logger <- layout_generator(msg_format = '{node}/{pid}/{namespace}/{fn} {time} {level}: {msg}')
logger(FATAL, 'try {runif(1)}')

log_layout(logger)
log_info('try {runif(1)}')
}
}
