% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_generator}
\alias{layout_generator}
\title{Generating logging function}
\usage{
layout_generator(msg_format = "{level} [{time}] {msg}",
  time_format = "\%Y-\%d-\%m \%H:\%M:\%S")
}
\arguments{
\item{msg_format}{\code{glue}-flavored layout of the message}

\item{time_format}{see \code{strptime} for details}

\item{level}{log level, eg \code{INFO}}

\item{msg}{character vector}
}
\value{
function taking \code{level} and \code{msg} arguments
}
\description{
Available variables to be used in the \code{msg_format}:
\itemize{
  \item level: log level, eg INFO
  \item time: current time formatted as \code{time_format}
  \item namespace: R package calling the logging function
  \item user: name of the real user id as reported by \code{Sys.info}
  \item pid: the process identification number of the R session
  \item node: name by which the machine is known on the network as reported by \code{Sys.info}
}
}
\examples{
\dontrun{
logger <- layout_generator(msg_format = '{node}/{pid} {time} {level}: {msg}')
logger(FATAL, 'asdsa {runif(1)}')
}
}
