% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_glue}
\alias{layout_glue}
\title{Format a log message with \code{glue}}
\usage{
layout_glue(level, msg, .call = sys.call(-1), .envir = parent.frame(1))
}
\arguments{
\item{level}{log level, see \code{\link{log_levels}} for more details}

\item{msg}{string message}
}
\value{
character vector
}
\description{
By default, this layout includes the log level of the log record as per \code{\link{log_levels}}, the current timestamp and the actual log message -- that you can override via calling \code{\link{layout_glue_generator}} directly. For colorized output, see \code{\link{layout_glue_colors}}.
}
\seealso{
This is a \code{\link{log_layout}}, for alternatives, see \code{\link{layout_simple}}, \code{\link{layout_glue_colors}}, \code{\link{layout_json}}, or generator functions such as \code{\link{layout_glue_generator}}
}
