% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_glue}
\alias{layout_glue}
\title{Formats a log message with \code{glue}}
\usage{
layout_glue(level, msg)
}
\arguments{
\item{level}{log level, eg \code{INFO}}

\item{msg}{character vector}
}
\value{
character vector
}
\description{
Formats a log message with \code{glue}
}
