% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_glue_colors}
\alias{layout_glue_colors}
\title{Format a log message with \code{glue} and ANSI escape codes to add colors}
\usage{
layout_glue_colors(level, msg)
}
\arguments{
\item{level}{log level, see \code{\link{log_levels}} for more details}

\item{msg}{string message}
}
\value{
character vector
}
\description{
Format a log message with \code{glue} and ANSI escape codes to add colors
}
