% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_json}
\alias{layout_json}
\title{Format a log message as JSON}
\usage{
layout_json(level, msg, .call = sys.call(-1), .envir = parent.frame())
}
\arguments{
\item{level}{log level, see \code{\link{log_levels}} for more details}

\item{msg}{string message}

\item{.call}{R expression from which the logging function was called (optionally used by the logging layout)}

\item{.envir}{original frame of the \code{.call} calling function where the formatter function is to be evaluated and that is used to look up the \code{namespace} as well via \code{\link{topenv}}}
}
\value{
character vector
}
\description{
Format a log message as JSON
}
\note{
This functionality depends on the \pkg{jsonlite} package.
}
\examples{
\dontrun{
log_layout(layout_json)
log_info(42:44)
}
}
\seealso{
This is a \code{\link{log_layout}}, for alternatives, see \code{\link{layout_simple}}, \code{\link{layout_glue}}, \code{\link{layout_glue_colors}} or generator functions such as \code{\link{layout_glue_generator}}
}
