% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_json}
\alias{layout_json}
\title{Format a log message as JSON}
\usage{
layout_json(level, msg)
}
\arguments{
\item{level}{log level, eg \code{INFO}}

\item{msg}{character vector}
}
\value{
character vector
}
\description{
Format a log message as JSON
}
\examples{
\dontrun{
log_layout(layout_json)
log_info(42:44)
}
}
