% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_json}
\alias{layout_json}
\title{Generate log layout function rendering JSON}
\usage{
layout_json(
  fields = c("time", "level", "ns", "ans", "topenv", "fn", "node", "arch", "os_name",
    "os_release", "os_version", "pid", "user")
)
}
\arguments{
\item{fields}{character vector of field names to be included in the
JSON}
}
\value{
character vector
}
\description{
Generate log layout function rendering JSON
}
\note{
This functionality depends on the \pkg{jsonlite} package.

This functionality depends on the \pkg{jsonlite} package.
}
\examples{
\dontrun{
log_layout(layout_json())
log_info(42)
log_info("ok {1:3} + {1:3} = {2*(1:3)}")
}
}
\seealso{
This is a \code{\link[=log_layout]{log_layout()}}, for alternatives, see
\code{\link[=layout_blank]{layout_blank()}}, \code{\link[=layout_simple]{layout_simple()}}, \code{\link[=layout_glue]{layout_glue()}},
\code{\link[=layout_glue_colors]{layout_glue_colors()}}, \code{\link[=layout_json_parser]{layout_json_parser()}}, or generator
functions such as \code{\link[=layout_glue_generator]{layout_glue_generator()}}

This is a \code{\link[=log_layout]{log_layout()}}, for alternatives, see \code{\link[=layout_blank]{layout_blank()}}, \code{\link[=layout_simple]{layout_simple()}}, \code{\link[=layout_glue]{layout_glue()}}, \code{\link[=layout_glue_colors]{layout_glue_colors()}}, \code{\link[=layout_json_parser]{layout_json_parser()}},  or generator functions such as \code{\link[=layout_glue_generator]{layout_glue_generator()}}
}
