% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_simple}
\alias{layout_simple}
\title{Format a log record by concatenating the log level, timestamp and message}
\usage{
layout_simple(level, msg, namespace = NA_character_,
  .call = sys.call(-1), .envir = parent.frame())
}
\arguments{
\item{level}{log level, see \code{\link{log_levels}} for more details}

\item{msg}{string message}

\item{namespace}{string referring to the \code{logger} environment / config to be used to override the target of the message record to be used instead of the default namespace, which is defined by the R package name from which the logger was called, and falls back to a common, global namespace.}

\item{.call}{R expression from which the logging function was called (optionally used by the logging layout)}

\item{.envir}{original frame of the \code{.call} calling function where the formatter function is to be evaluated and that is used to look up the \code{namespace} as well via \code{\link{topenv}}}
}
\value{
character vector
}
\description{
Format a log record by concatenating the log level, timestamp and message
}
\seealso{
This is a \code{\link{log_layout}}, for alternatives, see \code{\link{layout_glue}}, \code{\link{layout_glue_colors}}, \code{\link{layout_json}}, or generator functions such as \code{\link{layout_glue_generator}}
}
