% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouts.R
\name{layout_simple}
\alias{layout_simple}
\title{Format a log record by concatenating the log level, timestamp and message}
\usage{
layout_simple(level, msg)
}
\arguments{
\item{level}{log level, see \code{\link{log_levels}} for more details}

\item{msg}{string message}
}
\value{
character vector
}
\description{
Format a log record by concatenating the log level, timestamp and message
}
\seealso{
This is a \code{\link{log_layout}}, for alternatives, see \code{\link{layout_glue}}, \code{\link{layout_glue_colors}}, \code{\link{layout_json}}, or generator functions such as \code{\link{layout_glue_generator}}
}
