% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log}
\alias{log}
\title{Log a message with given log level}
\usage{
log(level, msg)
}
\arguments{
\item{level}{log level from \code{log_levels}}

\item{msg}{character vector}
}
\description{
Log a message with given log level
}
