% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log}
\alias{log}
\alias{log_fatal}
\alias{log_error}
\alias{log_warn}
\alias{log_info}
\alias{log_debug}
\alias{log_trace}
\title{Log a message with given log level}
\usage{
log(level, msg)
}
\arguments{
\item{level}{log level from \code{log_levels}}

\item{msg}{character vector}
}
\description{
Log a message with given log level
}
\examples{
\dontrun{
log(INFO, 'hi there')
log_info('hi there')

## output omitted
log_debug('hi there')

## TODO set threshold
}
}
