% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_appender}
\alias{log_appender}
\title{Get or set logger appender function}
\usage{
log_appender(appender, namespace = "global")
}
\arguments{
\item{namespace}{logger namespace}

\item{layout}{function}
}
\description{
Get or set logger appender function
}
\examples{
\dontrun{
t <- tempfile()
log_appender(appender_tee(t))
log_info(42)
log_info(42:44)
readLines(t)
}
}
