% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_formatter}
\alias{log_formatter}
\title{Get or set logger layout}
\usage{
log_formatter(formatter, namespace = "global", index = 1)
}
\arguments{
\item{formatter}{function defining how R objects are converted into a single string}

\item{namespace}{logger namespace}

\item{index}{index of the logger within the namespace}
}
\description{
Get or set logger layout
}
