% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_layout}
\alias{log_layout}
\title{Get or set log record layout}
\usage{
log_layout(layout = NULL, namespace = "global", index = 1)
}
\arguments{
\item{layout}{function defining the structure of a log record, eg
\code{\link[=layout_simple]{layout_simple()}}, \code{\link[=layout_glue]{layout_glue()}} or \code{\link[=layout_glue_colors]{layout_glue_colors()}},
\code{\link[=layout_json]{layout_json()}}, or generator functions such as
\code{\link[=layout_glue_generator]{layout_glue_generator()}}, default NULL}

\item{namespace}{logger namespace}

\item{index}{index of the logger within the namespace}
}
\description{
Get or set log record layout
}
\examples{
\dontshow{old <- logger:::namespaces_set()}
log_layout(layout_json())
log_info(42)
\dontshow{logger:::namespaces_set(old)}
}
\seealso{
Other log configutation functions: 
\code{\link{log_appender}()},
\code{\link{log_formatter}()},
\code{\link{log_threshold}()}
}
\concept{log configutation functions}
