% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_layout}
\alias{log_layout}
\title{Get or set logger layout}
\usage{
log_layout(layout, namespace = "global")
}
\arguments{
\item{layout}{function}

\item{namespace}{logger namespace}
}
\description{
Get or set logger layout
}
\examples{
\dontrun{
log_layout(layout_json)
log_info(42:44)
}
}
