% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_layout}
\alias{log_layout}
\title{Get or set log record layout}
\usage{
log_layout(layout, namespace = "global", index = 1)
}
\arguments{
\item{layout}{function defining the structure of a log record, eg \code{\link{layout_simple}}, \code{\link{layout_glue}} or \code{\link{layout_glue_colors}}, \code{\link{layout_json}}, or generator functions such as \code{\link{layout_glue_generator}}}

\item{namespace}{logger namespace}

\item{index}{index of the logger within the namespace}
}
\description{
Get or set log record layout
}
\examples{
\dontrun{
log_layout(layout_json())
log_info(42)
}
}
\seealso{
\code{\link{logger}}, \code{\link{log_threshold}}, \code{\link{log_appender}} and \code{\link{log_formatter}}
}
