% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\docType{data}
\name{FATAL}
\alias{FATAL}
\alias{log_levels}
\alias{ERROR}
\alias{WARN}
\alias{INFO}
\alias{DEBUG}
\alias{TRACE}
\title{Log levels}
\format{An object of class \code{loglevel} (inherits from \code{integer}) of length 1.}
\usage{
FATAL
}
\description{
Apache logj4 log levels:
}
\details{
\enumerate{
  \item FATAL
  \item ERROR
  \item WARN
  \item INFO
  \item DEBUG
  \item TRACE
}
}
\references{
\url{https://logging.apache.org/log4j/2.0/log4j-api/apidocs/org/apache/logging/log4j/Level.html}
}
\keyword{datasets}
