% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_separator}
\alias{log_separator}
\title{Logs a long line to stand out from the console}
\usage{
log_separator(level = INFO, separator = "=")
}
\arguments{
\item{level}{log level}

\item{separator}{character}
}
\description{
Logs a long line to stand out from the console
}
\examples{
log_separator()
log_separator(ERROR, '!')
}
