% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_threshold}
\alias{log_threshold}
\title{Get or set log level threshold}
\usage{
log_threshold(level, namespace = "global")
}
\arguments{
\item{level}{see \code{log_levels}}

\item{namespace}{logger namespace}
}
\value{
currently set log level threshold
}
\description{
Get or set log level threshold
}
