% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{log_warnings}
\alias{log_warnings}
\title{Injects a logger call to standard warnings}
\usage{
log_warnings()
}
\description{
This function uses \code{trace} to add a \code{log_warn} function call when \code{warning} is called to log the warning messages with the \code{logger} layout and appender.
}
\examples{
\dontrun{
log_warnings()
for (i in 1:5) { Sys.sleep(runif(1)); warning(i) }
}
}
