% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggeR-package.R
\docType{package}
\name{loggeR-package}
\alias{loggeR}
\alias{loggeR-package}
\title{loggeR: Basic logger utility}
\description{
Based on Logger utility from RestRserve package.
}
\author{
\strong{Maintainer}: Abraham Jaimes \email{abraham.jaimes.mx@gmail.com}

}
\keyword{internal}
