\docType{package}
\name{logger-package}
\alias{logger-package}
\title{Simple logger}
\description{
  Simple logger, logging using org-mode style headers and
  timestamps
}
\details{
  \tabular{ll}{ Package: \tab logger\cr Type: \tab
  Package\cr Version: \tab 0.0.1\cr Date: \tab
  2013-10-03_11-40\cr License: \tab GPL (>= 2)\cr LazyLoad:
  \tab yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

