% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Generate logging utility}
\usage{
logger(threshold, layout, appender)
}
\arguments{
\item{threshold}{omit log messages below this \code{log_levels}}

\item{layout}{function rendering the log message}

\item{appender}{function writing the log message}
}
\value{
function taking \code{level} and \code{msg} arguments
}
\description{
Generate logging utility
}
