% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Generate logging utility}
\usage{
logger(threshold, formatter, layout, appender)
}
\arguments{
\item{threshold}{omit log messages below this \code{\link{log_levels}}}

\item{formatter}{function pre-processing the message of the log record when it's not a single string with the \code{skip_formatter} attribute set to \code{TRUE}}

\item{layout}{function rendering the layout of the actual log record}

\item{appender}{function writing the log record}
}
\value{
function taking \code{level} and \code{msg} arguments
}
\description{
A logger consists of a log level \code{threshold}, a log message \code{formatter} function, a log record \code{layout} formatting function and the \code{appender} function deciding on the destination of the log record. For more details, see the package \code{README.md}.
}
\note{
It's quite unlikely that you need to call this function directly, but instead set the logger parameters and functions at \code{\link{log_threshold}}, \code{\link{log_formatter}}, \code{\link{log_layout}} and \code{\link{log_appender}} and then call \code{\link{log_levels}} and its derivatives, such as \code{\link{log_info}} directly.
}
