% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Class to provide logging capabilities}
\description{
Class to provide logging capabilities

Class to provide logging capabilities
}
\examples{

## ------------------------------------------------
## Method `logger$new`
## ------------------------------------------------

##create a new log
log1 <- logger$new()
log2 <- logger$new("log1.log")

## ------------------------------------------------
## Method `logger$add`
## ------------------------------------------------

## add a log record
log1$add(log_lvl = "debug", log_msg = "some information")
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{file}}{Set or get the file path of the logger}

\item{\code{custom_sanitizer}}{get, replace or add custom sanitizer rules to the default sanitizer}

\item{\code{default_sanitizer}}{get the default sanitizer rules}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{logger$new()}}
\item \href{#method-add}{\code{logger$add()}}
\item \href{#method-warn}{\code{logger$warn()}}
\item \href{#method-info}{\code{logger$info()}}
\item \href{#method-err}{\code{logger$err()}}
\item \href{#method-return_log}{\code{logger$return_log()}}
\item \href{#method-clone}{\code{logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize a new logger with \code{logger$new("file_path")}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$new(file = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file path to log file that already exists or should be created (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{##create a new log
log1 <- logger$new()
log2 <- logger$new("log1.log")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
add a log entry to the logging file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$add(log_lvl, log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_lvl}}{character string for the log level to record the log entry for}

\item{\code{log_msg}}{log message to record}

\item{\code{...}}{custom named values to record in the log}

\item{\code{echo}}{Overriding echo = T argument}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{## add a log record
log1$add(log_lvl = "debug", log_msg = "some information")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-warn"></a>}}
\if{latex}{\out{\hypertarget{method-warn}{}}}
\subsection{Method \code{warn()}}{
add a log entry with "warn" as the \code{log_lvl}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$warn(log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_msg}}{log message to record}

\item{\code{...}}{custom named values to record in the log}

\item{\code{echo}}{Overriding echo = T argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
add a log entry with "info" as the \code{log_lvl}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$info(log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_msg}}{log message to record}

\item{\code{...}}{custom named values to record in the log}

\item{\code{echo}}{Overriding echo = T argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-err"></a>}}
\if{latex}{\out{\hypertarget{method-err}{}}}
\subsection{Method \code{err()}}{
add a log entry with "err" as the \code{log_lvl}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$err(log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{log_msg}}{log message to record}

\item{\code{...}}{custom named values to record in the log}

\item{\code{echo}}{Overriding echo = T argument}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_log"></a>}}
\if{latex}{\out{\hypertarget{method-return_log}{}}}
\subsection{Method \code{return_log()}}{
return the log as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$return_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
