% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{logger}
\alias{logger}
\title{Class to provide logging capabilities}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods for logging events to a log file.
}
\description{
Class to provide logging capabilities

Class to provide logging capabilities
}
\examples{
my_log <- logger$new("log1.log")
my_log$info("This is an information")
}
\keyword{logger}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{file}}{stores the file path}

\item{\code{custom_sanitizer}}{stores the customer sanitizer list which cleanes logging strings for ndjson format.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{logger$new()}}
\item \href{#method-add}{\code{logger$add()}}
\item \href{#method-warn}{\code{logger$warn()}}
\item \href{#method-info}{\code{logger$info()}}
\item \href{#method-err}{\code{logger$err()}}
\item \href{#method-return_log}{\code{logger$return_log()}}
\item \href{#method-clone}{\code{logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$new(file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$add(log_lvl, log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-warn"></a>}}
\if{latex}{\out{\hypertarget{method-warn}{}}}
\subsection{Method \code{warn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$warn(log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-info"></a>}}
\if{latex}{\out{\hypertarget{method-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$info(log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-err"></a>}}
\if{latex}{\out{\hypertarget{method-err}{}}}
\subsection{Method \code{err()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$err(log_msg, ..., echo = T)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-return_log"></a>}}
\if{latex}{\out{\hypertarget{method-return_log}{}}}
\subsection{Method \code{return_log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$return_log()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
