% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loggerfiles.R
\name{read_loggerfiles}
\alias{read_loggerfiles}
\title{Read and combine logger csv files}
\usage{
read_loggerfiles(
  loggerinfo,
  add_info = c("SN", "type", "site", "habitat", "replicate"),
  folder = "data/raw_csv",
  check_SN = TRUE
)
}
\arguments{
\item{loggerinfo}{The name of the `dataframe`containing the logger metadata}

\item{add_info}{a character vector of column names in `loggerinfo`. The content of these columns will be added to the corresponding logger data}

\item{folder}{The path to the folder containing the logger `.csv` files. The default is `data/raw_csv`}

\item{check_SN}{logical: Should the serial number (SN) provided in the metadata dataframe be checked against the SN stored in the logger `.csv` files? The default is `TRUE`.}
}
\value{
A `dataframe` in "long" format containing all the logger data. Additional information, as SN, filename, logger type, etc, is added
}
\description{
Uses the info provided in `loggerinfo` to read and merge all logger csv files.
}
\examples{
\dontrun{
dat_logger <- read_loggerfiles(loggerinfo)
}

}
